% Generated by roxygen2 (4.0.2): do not edit by hand
\name{i18n}
\alias{i18n}
\title{Translate parts of a plugin}
\usage{
i18n(..., lang = rk.get.language())
}
\arguments{
\item{...}{Comma separated, named elements, see description.}

\item{lang}{Character string, the language to return.}
}
\description{
Takes a list of entries named after abbreviated languages, and returns
either the one matching the language set with \code{\link[rkwarddev:rk.set.language]{rk.set.language}},
or the first entry if no language was set at all or the set language cannot be found in
\code{...}.
}
\details{
If used in an \code{rkwarddev} script, this can be used to toggle the generation of plugins
in a certain language.
}
\examples{
rk.set.language("en", c("en_EN", "en_US"))
(var.select <- rk.XML.varselector(label=i18n(en="Select data", de="Wähle Daten")))

# now try the same with the alternate language
rk.set.language("de", "de_DE")
(var.select <- rk.XML.varselector(label=i18n(en="Select data", de="Wähle Daten")))
}

