% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifiers}
\alias{modifiers}
\title{Get all valid modifiers for a given XML node}
\usage{
modifiers(obj = "all")
}
\arguments{
\item{obj}{An object of class \code{XiMpLe.node}, or a character string containing
the XML node name you're interested in. If set to \code{"all"} returns all defined
modifiers.}
}
\value{
A character vector of valid modifiers.
}
\description{
In case you want to see which modifiers are definied for a certain XML node,
just call this helper function.
}
\examples{
myCheckbox <- rk.XML.cbox("Check for action")
modifiers(myCheckbox)
}

