% Generated by roxygen2 (4.0.2): do not edit by hand
\name{R.comment}
\alias{R.comment}
\title{Create R comment with JavaScript in RKWard plugin code}
\usage{
R.comment(message, indent.by = "\\\\t", level = 2, newline = "")
}
\arguments{
\item{message}{}

\item{indent.by}{A character string defining the indentation string to use. Note that
backslashes need to be escaped (e.g. \code{"\\t"} to produce \code{"\t"}).}

\item{level}{Integer, which indentation level to use in the resulting R code,
      minimum is 1.}

\item{newline}{Character string,
      can be set to e.g. \code{"\n"} to force a newline after the call.}
}
\value{
A character string.
}
\description{
Create R comment with JavaScript in RKWard plugin code
}
\examples{
cat(R.comment("This will become an R comment"))
}

