% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rk.get.comp}
\alias{rk.get.comp}
\title{Get the name of the component you're currently working on}
\usage{
rk.get.comp()
}
\description{
This is the "get"-equivalent to \code{\link[rkwarddev:rk.set.comp]{rk.set.comp}}.
It is used by functions like, e.g., \code{\link[rkwarddev:rk.XML.cbox]{rk.XML.cbox}},
to add text for .rkh pages automatically to the current plugin component.
}

