% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rk.get.language}
\alias{rk.get.language}
\title{Get plugin language for internationalisation}
\usage{
rk.get.language(locales = FALSE)
}
\arguments{
\item{locales}{Logical, whether to query for language or locales set.}
}
\description{
Get plugin language for internationalisation
}
\examples{
rk.get.language()
}

