% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.i18n.comment.R
\name{rk.i18n.comment}
\alias{rk.i18n.comment}
\title{Create i18n comment for RKWard plugin code}
\usage{
rk.i18n.comment(text, prefix = "i18n:")
}
\arguments{
\item{text}{Character string, the text to be displayed.}

\item{prefix}{Character string, the text to be prefixed to indicate this is
an i18n comment.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
This function is similar to rk.comment, but preceds the \code{text} with the
keyword \code{"i18n:"} to give context to translators.
}
\examples{
test.comment <- rk.i18n.comment("Added this text.")
cat(pasteXML(test.comment))
}

