% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_01_rk.JS.arr.R
\docType{class}
\name{rk.JS.arr-class}
\alias{rk.JS.arr-class}
\title{S4 Class rk.JS.arr}
\description{
This simple class is used for JavaScript generation and produced by
\code{\link[rkwarddev:rk.JS.array]{rk.JS.array}}. You shouldn't
need to temper with this type of class manually.
}
\section{Slots}{

\describe{
\item{\code{arr.name}}{Character string, name of the array variable.}

\item{\code{opt.name}}{Character string, name of the option variable.}

\item{\code{IDs}}{Character vector of IDs.}

\item{\code{variables}}{Character vector of variables.}

\item{\code{funct}}{Character string, name of an R function call.}

\item{\code{quote}}{Logical, should values be quoted?}

\item{\code{option}}{Character string, name of the option to set.}

\item{\code{opt.sep}}{Character string, separates previous options from the one defined by the array.}
}}
\keyword{Classes}

