% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/00_class_03_rk.JS.ite.R
\docType{class}
\name{rk.JS.ite-class}
\alias{rk.JS.ite-class}
\title{S4 Class rk.JS.ite}
\description{
This simple class is used for JavaScript generation and produced by
\code{\link{ite}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{ifJS}}{Character string, the "if" clause.}

\item{\code{thenJS}}{Character string, the "then" body.}

\item{\code{thenifJS}}{A list with exactly one optional object of class rk.JS.ite.}

\item{\code{elseJS}}{Character string, the "else" body.}

\item{\code{elifJS}}{A list with exactly one optional object of class rk.JS.ite.}
}}
\keyword{Classes}

