% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.JS.method.R
\name{rk.JS.method}
\alias{rk.JS.method}
\title{Simple JavaScript method generator}
\usage{
rk.JS.method(name, values = NULL, suffix = NULL, object = NULL,
  var = NULL)
}
\arguments{
\item{name}{Character string, name of the mthod to call on the JS object.}

\item{values}{Character vector, argument values for the method.}

\item{suffix}{Character string, will be appended to the foll call as-is.}

\item{object}{Either a character string (name of the JS object), a XiMpLe node (whose
ID will be used), or an object of class \code{rk.JS.var} (whose JS name will be used).}

\item{var}{Character string, name of a JS variable to define. This will also append
";" at the end of the generated string.}
}
\value{
A character string.
}
\description{
Combines argument values into a JS method to append to an object call.
}
\examples{
rk.JS.method("split", values="[[", suffix="[0]")
}

