% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_04_rk.JS.opt.R
\docType{class}
\name{rk.JS.opt-class}
\alias{rk.JS.opt-class}
\title{S4 Class rk.JS.opt}
\description{
This simple class is used for JavaScript generation and produced by
\code{\link{rk.JS.options}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{var.name}}{Character string, the name of the variable.}

\item{\code{opt.name}}{Character string, the name of the option.}

\item{\code{collapse}}{Character string,
      used to collapse several options into one string.}

\item{\code{ifs}}{A list with objects of class rk.JS.ite.}

\item{\code{array}}{Logical, whether to use an array for options.}

\item{\code{funct}}{Character string,
      name of the R function to be called to combine the options.}

\item{\code{opt.sep}}{Character string,
      separates previous options from the one defined here.}
}}
\keyword{Classes}

