% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_05_rk.JS.oset.R
\docType{class}
\name{rk.JS.oset-class}
\alias{rk.JS.oset-class}
\title{S4 Class rk.JS.oset}
\description{
This simple class is used for JavaScript generation and produced by
\code{\link{rk.JS.optionset}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{vars}}{An object of class \code{rk.JS.var}.}

\item{\code{loopvar}}{Character string, name of the index variable used in the for loop.}

\item{\code{columns}}{A list of <optioncolumn> nodes.}

\item{\code{body}}{A list of JavaScript code, the body of the for loop.}

\item{\code{collapse}}{Character string,
      how all optioncolumns should be concatenated on the R code level.}
}}
\keyword{Classes}

