% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_02_rk.JS.var.R
\docType{class}
\name{rk.JS.var-class}
\alias{rk.JS.var-class}
\title{S4 Class rk.JS.var}
\description{
This simple class is used for JavaScript generation and produced by
\code{\link{rk.JS.vars}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{JS.var}}{Character string, name of the JavaScript variable.}

\item{\code{XML.var}}{Character string, name of the XML variable.}

\item{\code{prefix}}{Character string, an optional prefix for variable names.}

\item{\code{modifiers}}{A list of modifiers to apply to the XML node property.}

\item{\code{default}}{Logical,
      whether the default value (no special modifier) of the node should also be defined.}

\item{\code{append.modifier}}{Logical, if a modifier is given,
      should that become part of the variable name?}

\item{\code{join}}{Character string,
      if set is used to collapse multiple values into one string.}

\item{\code{vars}}{A list of objects of class rk.JS.var.}

\item{\code{getter}}{Character string,
      the JavaScript function which should be used to fetch the values from the plugin.}
}}
\keyword{Classes}

