% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.local.R
\name{rk.local}
\alias{rk.local}
\title{Evaluate code in a local environment}
\usage{
rk.local(...)
}
\arguments{
\item{...}{The code to be evaluated.}
}
\value{
The result of evaluating the object(s).
}
\description{
Can be used like \code{\link[base:local]{local}}, but evaluation is being done in a speacial
local environment of the rkwarddev package. This can be neccessary if you want to call functions
nested insinde \code{\link[rkwarddev:js]{js}}, because it might not find all objects if they were
only defined in a standard local environment.
}

