\name{rk.paste.JS.graph}
\alias{rk.paste.JS.graph}
\title{Paste simple JavaScript plot code}
\usage{
  rk.paste.JS.graph(..., plotOpts = NULL,
    printoutObj = NULL, level = 2, indent.by = "\t",
    empty.e = FALSE)
}
\arguments{
  \item{...}{The actual plot code, passed through to
  \code{rk.paste.JS}.}

  \item{plotOpts}{A node object generated by
  \code{rk.XML.embed}, i.e. embedded plot options.}

  \item{printoutObj}{An \code{rk.JS.var} object fetching
  the \code{"code.printout"} modifier of \code{plotOpts}
  (see examples below!).}
}
\value{
  A character string.
}
\description{
  This function is similar to \code{rk.paste.JS}, but adds
  some code parts to its output which are commonly used to
  generate plots with RKWard.
}
\details{
  The contents of the \code{...} argument are evaluated by
  \code{rk.paste.JS} and encapsulated between
  \code{if(full)\{rk.graph.on()\} try(\{} and \code{\})
  if(full)\{rk.graph.off()\}}. If generic plot options are
  supplied, their \code{"code.preprocess"} and
  \code{"code.calculate"} modifiers are also automatically
  taken care of, so you only need to include
  \code{"code.printout"} inside of \code{...}.
}
\examples{
tmp.var.selectVars <- rk.XML.varselector(label="Select data")
tmp.var.x <- rk.XML.varslot(label="My data", source=tmp.var.selectVars, required=TRUE)
# let this be the embedded generic plot options in yout plot dialog
tmp.plot.options <- rk.XML.embed(component="rkward::plot_options", button=TRUE, label="Generic plot options")

# now generate an rk.JS.var object to import the "printout" code from the embedded options
js.po.printout <- rk.JS.vars(tmp.plot.options, modifiers="code.printout", check.modifiers=FALSE)

# you can now generate the plot code using generic plot options
js.prnt <- rk.paste.JS.graph(
	echo("\\t\\tplot("),
	echo("\\n\\t\\t\\tx=", tmp.var.x),
	echo(js.po.printout),
	echo(")"),
	plotOpts=tmp.plot.options,
	printoutObj=js.po.printout)

cat(js.prnt)
}
\seealso{
  \code{\link[rkwarddev:rk.paste.JS]{rk.paste.JS}}
}

