% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.rk.plot.opts-class.R
\docType{class}
\name{rk.plot.opts-class}
\alias{rk.plot.opts-class}
\title{S4 Class rk.plot.opts}
\description{
This simple class is used for JavaScript generation and is produced by
\code{\link{rk.plotOptions}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{XML}}{An object of class XiMpLe.node.}

\item{\code{preprocess}}{An object of class rk.JS.var.}

\item{\code{printout}}{An object of class rk.JS.var.}

\item{\code{calculate}}{An object of class rk.JS.var.}
}}
\keyword{Classes}

