% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{rk.plug.comp-class}
\alias{rk.plug.comp-class}
\title{S4 Class rk.plug.comp}
\description{
This simple class is used for JavaScript generation. It holds plugin components,
i.e. single dialogs, to add to a plugin skeleton, and is produced by
\code{\link{rk.plugin.component}}. You shouldn't need to temper with this
type of class manually.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character string, name of the plugin.}

\item{\code{create}}{Charactervector defining the component parts/files to be created.}

\item{\code{xml}}{An object of class XiMpLe.doc containig the plugin XML code.
See \code{\link{rk.XML.plugin}}.}

\item{\code{js}}{A character string containing the plugin JavaScript code.
See \code{\link{rk.JS.doc}}.}

\item{\code{rkh}}{An object of class XiMpLe.doc containig the plugin help page.
See \code{\link{rk.rkh.doc}}.}

\item{\code{hierarchy}}{A list defining where to place the component in the menu structure.}
}}
\keyword{Classes}

