% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.set.comp.R
\name{rk.set.comp}
\alias{rk.set.comp}
\title{Define the component you're currently working on}
\usage{
rk.set.comp(component = NULL)
}
\arguments{
\item{component}{Character string, name of the component to set. If \code{NULL},
no component will be set as default, and \code{rk.get.comp} will return \code{NULL}
subsequently.}
}
\description{
This small tool let's you set a component name as kind of "active", which simply
means it will be returned by \code{\link[rkwarddev:rk.get.comp]{rk.get.comp}}. This can be
used by functions like, e.g., \code{\link[rkwarddev:rk.XML.cbox]{rk.XML.cbox}}, to add text
for .rkh pages automatically to the current plugin component.
}

