% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.get.empty.e.R, R/rk.set.empty.e.R
\name{rk.get.empty.e}
\alias{rk.get.empty.e}
\alias{rk.set.empty.e}
\title{Globally define handling of empty else conditions in JS}
\usage{
rk.get.empty.e()

rk.set.empty.e(empty = FALSE)
}
\arguments{
\item{empty}{Logical, whether .}
}
\value{
\code{rk.set.empty.e} returns invisible(NULL),
      \code{rk.get.empty.e} either \code{TRUE} or \code{FALSE}.
}
\description{
Some JS functions allow to decide whether empty \code{else} statements should be omitted or printed
nonetheless (which some consider more reader friendly). The default can be globally defined with
\code{rk.set.empty.e}, so you don't have to specify it in each function call.
}
\details{
\code{rk.get.empty.e} returns the set value, which defaults to \code{FALSE} by default.
}

