% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rk.set.language}
\alias{rk.set.language}
\title{Set plugin language for internationalisation}
\usage{
rk.set.language(lang = NULL, locales = NULL)
}
\arguments{
\item{lang}{Character string, abbreviated language to use with \code{\link[rkwarddev:i18n]{i18n}}, e.g. \code{"en"}.}

\item{locales}{Character vector, all the locales this translation covers, e.g. \code{c("en_EN", "en_US")}.}
}
\description{
Stores the given language in an internal environment, so functions like
\code{\link[rkwarddev:i18n]{i18n}} can use it.
}
\examples{
rk.set.language("en", c("en_EN", "en_US"))
}

