% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{show,-methods}
\alias{show,rk.JS.arr-method}
\alias{show,rk.JS.echo-method}
\alias{show,rk.JS.ite-method}
\alias{show,rk.JS.opt-method}
\alias{show,rk.JS.var-method}
\title{Show methods for objects of class rk.JS.S}
\usage{
show(object)

\S4method{show}{rk.JS.arr}(object)

\S4method{show}{rk.JS.ite}(object)

\S4method{show}{rk.JS.opt}(object)

\S4method{show}{rk.JS.var}(object)

\S4method{show}{rk.JS.echo}(object)
}
\arguments{
\item{object}{An object of class \code{rk.JS.*}}
}
\description{
Show methods for S4 objects of class \code{rk.JS.*}
}
\keyword{methods}

