\name{rk.paste.JS}
\alias{rk.paste.JS}
\title{Paste objects of class rk.JS.ite}
\usage{
  rk.paste.JS(object, level = 1, indent.by = "\t")
}
\arguments{
  \item{object}{An object of class \code{rk.JS.ite}}

  \item{level}{Integer, which indentation level to use,
  minimum is 1.}

  \item{indent.by}{A character string defining the
  indentation string to use.}
}
\value{
  A character string.
}
\description{
  Paste objects of class rk.JS.ite
}

