% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rk.rkh.scan}
\alias{rk.rkh.scan}
\title{Create RKWard help nodes from plugin XML}
\usage{
rk.rkh.scan(pXML, help = TRUE, captions = TRUE, component = NULL)
}
\arguments{
\item{pXML}{Either an object of class \code{XiMpLe.doc} or \code{XiMpLe.node},
      or path to a plugin XML file.}

\item{help}{Logical, if \code{TRUE} a list of XiMpLe.node objects will be returned,
      otherwise a character
vector with only the relevant ID names.}

\item{captions}{Logical, if \code{TRUE} captions will be generated for all "page",
      "tab" and "frame" nodes.}

\item{component}{Character string,
      name of the scanned component. Only needed if you want to search for
help text provided by \code{\link[rkwarddev:rk.set.rkh.prompter]{rk.set.rkh.prompter}}.}
}
\value{
A character vector or a list of XiMpLe.node objects. Returns \code{NULL} if no documentable nodes are found.
}
\description{
Create RKWard help nodes from plugin XML
}
\seealso{
\href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

