% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rk.rkh.technical.R
\name{rk.rkh.technical}
\alias{rk.rkh.technical}
\title{Create XML "technical" node for RKWard help pages}
\usage{
rk.rkh.technical(text = NULL, i18n = NULL)
}
\arguments{
\item{text}{Character string, the text to be displayed.}

\item{i18n}{Either a character string or a named list with the optional elements \code{context}
or \code{comment}, to give some \code{i18n_context} information for this node.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create XML "technical" node for RKWard help pages
}
\examples{
plugin.technical <- rk.rkh.technical("<b>TODO</b>: Implement sandworm detector.")
cat(pasteXML(plugin.technical))
}
\seealso{
\code{\link[rkwarddev:rk.rkh.doc]{rk.rkh.doc}}
   and the \href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

