% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rk.rkh.usage}
\alias{rk.rkh.usage}
\title{Create XML "usage" node for RKWard help pages}
\usage{
rk.rkh.usage(text = NULL, i18n = NULL)
}
\arguments{
\item{text}{Character string, the text to be displayed.}

\item{i18n}{Either a character string or a named list with the optional element \code{context},
to give some \code{i18n_context}
information for this node.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create XML "usage" node for RKWard help pages
}
\examples{
plugin.usage <- rk.rkh.usage("First do this, then do that ...")
}
\seealso{
\code{\link[rkwarddev:rk.rkh.doc]{rk.rkh.doc}}
   and the \href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

