\name{rk.testsuite.doc}
\alias{rk.testsuite.doc}
\title{Create testsuite outline to test an RKWard plugin}
\usage{
  rk.testsuite.doc(name = NULL)
}
\arguments{
  \item{name}{A character string, name of the
  plugin/dialog.}
}
\value{
  A character string.
}
\description{
  Create testsuite outline to test an RKWard plugin
}
\seealso{
  \href{help:rkwardplugins}{Introduction to Writing Plugins
  for RKWard}
}

