% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.uniqueIDs.R
\name{rk.uniqueIDs}
\alias{rk.uniqueIDs}
\title{Check plugin dialogs for duplicate IDs}
\usage{
rk.uniqueIDs(obj, bool = FALSE, warning = TRUE, ignore = c("copy"))
}
\arguments{
\item{obj}{An XML object of class XiMpLe.node or XiMpLe.doc.}

\item{bool}{Logical, if \code{TRUE} this function will return a logical value.}

\item{warning}{Logical, if \code{TRUE} will throw a warning if duplicates were found,
listing the findings.}

\item{ignore}{Character vector, node names that should generally be ignored because
they duplicate IDs by design.}
}
\value{
A vector with duplicate IDs, if any were found.
   If \code{bool=TRUE} returns a logical value.
}
\description{
A plugin must not have duplicated IDs to work properly. This function
cannot automatically correct duplicates, but it will warn you about
it, so you can correct your plugin script manually
}

