% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.XML.col.R
\name{rk.XML.col}
\alias{rk.XML.col}
\title{Create XML node "column" for RKWard plugins}
\usage{
rk.XML.col(..., id.name = "auto")
}
\arguments{
\item{...}{Objects of class \code{XiMpLe.node}.}

\item{id.name}{Character string, a unique ID for this plugin element.
If \code{"auto"}, an ID will be generated automatically from the objects in \code{...}.
If \code{NULL}, no ID will be given.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create XML node "column" for RKWard plugins
}
\examples{
test.checkboxes <- rk.XML.row(rk.XML.col(
    rk.XML.cbox(label="foo", val="foo1", chk=TRUE),
    rk.XML.cbox(label="bar", val="bar2")))
cat(pasteXML(test.checkboxes))
}
\seealso{
\href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

