\name{rk.XML.context}
\alias{rk.XML.context}
\title{Create XML "context" node for RKWard plugins}
\usage{
  rk.XML.context(id = "x11")
}
\arguments{
  \item{id}{Either "x11" or "import".}
}
\value{
  A list of objects of class \code{XiMpLe.node}.
}
\description{
  This function will create a context node for .pluginmap
  files.
}
\note{
  NOT WORKING YET
}

