\name{rk.XML.entry}
\alias{rk.XML.entry}
\title{Create XML "entry" node for RKWard plugins}
\usage{
  rk.XML.entry(component, index = 1)
}
\arguments{
  \item{component}{An ID.}

  \item{index}{An index.}
}
\value{
  A list of objects of class \code{XiMpLe.node}.
}
\description{
  This function will create a entry node for .pluginmap
  files.
}
\note{
  NOT WORKING YET
}
\seealso{
  \code{\link[rkwarddev:rk.XML.entrys]{rk.XML.entrys}}
}

