\name{rk.XML.hierarchy}
\alias{rk.XML.hierarchy}
\title{Create XML hierarchy section for RKWard plugins}
\usage{
  rk.XML.hierarchy(nodes)
}
\arguments{
  \item{nodes}{A (list of) objects of class
  \code{XiMpLe.node}.}
}
\value{
  A list of objects of class \code{XiMpLe.node}.
}
\description{
  This function will create a hierarchy section with
  optional child nodes "menu".
}
\note{
  NOT WORKING YET
}

