\name{rk.XML.input}
\alias{rk.XML.input}
\title{Create XML node "input" for RKWard plugins}
\usage{
  rk.XML.input(label, id.name = "auto")
}
\arguments{
  \item{label}{Character string, a text label for this
  plugin element.}

  \item{id.name}{Character string, a unique ID for this
  plugin element. If \code{"auto"}, an ID will be generated
  automatically from the label.}
}
\value{
  An object of class \code{XiMpLe.node}.
}
\description{
  Create XML node "input" for RKWard plugins
}

