% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rk.XML.option}
\alias{rk.XML.option}
\title{Create XML node "option" for RKWard plugins}
\usage{
rk.XML.option(label, val = NULL, chk = FALSE, id.name = NULL)
}
\arguments{
\item{label}{Character string, a text label for this plugin element.}

\item{val}{Character string, defines the value to submit if the option is checked.}

\item{chk}{Logical,
      should be set \code{TRUE} in the one option which is checked by default.}

\item{id.name}{Character string, a unique ID for this plugin element.
If \code{"auto"} and a label was provided,
      an ID will be generated automatically from the label.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create XML node "option" for RKWard plugins
}
\examples{
test.radio <- rk.XML.radio("Chose one",
  options=list(
    rk.XML.option("First Option", val="val1"),
    rk.XML.option("Second Option", val="val2", chk=TRUE))
cat(pasteXML(test.radio))
}
\seealso{
\href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

