% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rk.XML.optiondisplay}
\alias{rk.XML.optiondisplay}
\title{Create XML node "optiondisplay" for RKWard plugins}
\usage{
rk.XML.optiondisplay(index = TRUE, id.name = NULL)
}
\arguments{
\item{index}{Logical, whether to show a column with a numeric index in the optiondisplay.}

\item{id.name}{Character string, a unique ID for this plugin element (optional).}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
This node is only allowed once insinde the \code{<content>} node of an \code{<optionset>}.
}
\note{
The \code{<optionset>} node was introduced with RKWard 0.6.1, please set the dependencies
   of your component/plugin accordingly.
}
\seealso{
\code{\link[rkwarddev:rk.XML.optionset]{rk.XML.optionset}},
   \code{\link[rkwarddev:rk.XML.optioncolumn]{rk.XML.optioncolumn}},
   and the \href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

