\name{rk.XML.require}
\alias{rk.XML.require}
\title{Create XML "require" node for RKWard plugins}
\usage{
  rk.XML.require(file)
}
\arguments{
  \item{file}{A file name.}
}
\value{
  A list of objects of class \code{XiMpLe.node}.
}
\description{
  This function will create a require node for .pluginmap
  files.
}
\note{
  NOT WORKING YET
}

