\name{rk.XML.row}
\alias{rk.XML.row}
\title{Create XML node "row" for RKWard plugins}
\usage{
  rk.XML.row(children = list(), id.name = NULL)
}
\arguments{
  \item{children}{An optional list with objects of class
  \code{XiMpLe.node}.}

  \item{id.name}{Character string, a unique ID for this
  plugin element. If \code{NULL}, no ID will be given.}
}
\value{
  An object of class \code{XiMpLe.node}.
}
\description{
  Create XML node "row" for RKWard plugins
}

