% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rk.XML.saveobj.R
\name{rk.XML.saveobj}
\alias{rk.XML.saveobj}
\title{Create XML node "saveobject" for RKWard plugins}
\usage{
rk.XML.saveobj(label, chk = FALSE, checkable = TRUE, initial = "auto",
  required = FALSE, id.name = "auto", help = NULL,
  component = rk.get.comp(), i18n = NULL)
}
\arguments{
\item{label}{Character string, a text label for this plugin element.}

\item{chk}{Logical,
      if \code{TRUE} and \code{checkable=TRUE} the option is checkable and active by default.}

\item{checkable}{Logical, if \code{TRUE} the option can be switched on and off.}

\item{initial}{Character string, the default name for the object should be saved to.
If \code{"auto"} and a label was provided,
      an name will be generated automatically from the label.}

\item{required}{Logical, whether an entry is mandatory or not.}

\item{id.name}{Character string, a unique ID for this plugin element.
If \code{"auto"} and a label was provided,
      an ID will be generated automatically from the label.}

\item{help}{Character string or list of character values and XiMpLe nodes,
      will be used as the \code{text} value for a setting node in the .rkh file.
If set to \code{FALSE},
      \code{\link[rkwarddev:rk.rkh.scan]{rk.rkh.scan}} will ignore this node.
Also needs \code{component} to be set accordingly!}

\item{component}{Character string,
      name of the component this node belongs to. Only needed if you
want to use the scan features for automatic help file generation; needs \code{help} to be set
accordingly, too!}

\item{i18n}{Either a character string or a named list with the optional elements \code{context}
or \code{comment},
      to give some \code{i18n_context} information for this node. If set to \code{FALSE},
the attribute \code{label} will be renamed into \code{noi18n_label}.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create XML node "saveobject" for RKWard plugins
}
\examples{
test.saveobj <- rk.XML.saveobj("Save the results")
cat(pasteXML(test.saveobj))
}
\seealso{
\href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

