% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rk.XML.valueselector}
\alias{rk.XML.valueselector}
\title{Create node "valueselector" for RKWard plugins}
\usage{
rk.XML.valueselector(label = NULL, id.name = "auto")
}
\arguments{
\item{label}{Character string, a text label for the value selection slot.
Must be set if \code{id.name="auto"}.}

\item{id.name}{Character vector, unique ID for this element.}
}
\value{
An object of class \code{XiMpLe.node}.
}
\description{
Create node "valueselector" for RKWard plugins
}
\examples{
test.valueselector <- rk.XML.valueselector("Select some values")
cat(pasteXML(test.valueselector))
}
\seealso{
\code{\link[rkwarddev:rk.XML.valueslot]{rk.XML.valueslot}},
   \code{\link[rkwarddev:rk.XML.values]{rk.XML.values}},
   and the \href{help:rkwardplugins}{Introduction to Writing Plugins for RKWard}
}

