% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rkwarddev-package}
\alias{rkwarddev-package}
\title{The rkwarddev Package}
\description{
A collection of tools for RKWard plugin development.
}
\details{
\tabular{ll}{
Package: \tab rkwarddev\cr
Type: \tab Package\cr
Version: \tab 0.06-6\cr
Date: \tab 2014-12-17\cr
Depends: \tab R (>= 2.9.0),methods,XiMpLe (>= 0.03-21),rkward (>= 0.5.7)\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://rkward.kde.org\cr
}

Provides functions to create plugin skeletons and XML structures for RKWard.
}
\author{
m.eik michalke
}
\keyword{package}

