\docType{package}
\name{rkwarddev-package}
\alias{rkwarddev}
\alias{rkwarddev-package}
\title{A collection of tools for RKWard plugin development.}
\description{
  A collection of tools for RKWard plugin development.
}
\details{
  \tabular{ll}{ Package: \tab rkwarddev\cr Type: \tab
  Package\cr Version: \tab 0.04-1\cr Date: \tab
  2011-10-20\cr Depends: \tab R (>= 2.9.0),XiMpLe,rkward
  (>= 0.5.7)\cr Enhances: \tab rkward\cr Encoding: \tab
  UTF-8\cr License: \tab GPL (>= 3)\cr LazyLoad: \tab
  yes\cr URL: \tab http://rkward.sourceforge.net\cr }

  Provides functions to create plugin skeletons and XML
  structures for RKWard.
}
\author{
  m.eik michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

