% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkwarddev-package.R
\docType{package}
\name{rkwarddev-package}
\alias{rkwarddev}
\alias{rkwarddev-package}
\title{\packageTitle{rkwarddev}}
\description{
\packageDescription{rkwarddev}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab rkwarddev\cr
Type: \tab Package\cr
Version: \tab 0.08-3\cr
Date: \tab 2018-10-12\cr
Depends: \tab R (>= 2.9.0),XiMpLe (>= 0.03-23),rkward (>= 0.5.7)\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
  \item Report bugs at \url{https://github.com/rkward-community/rkwarddev/issues}
}

}
\author{
\packageAuthor{rkwarddev}

Maintainer: \packageMaintainer{rkwarddev}
}
