% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rkwarddev.required.R
\name{rkwarddev.required}
\alias{rkwarddev.required}
\title{Check for rkwarddev package version requirements}
\usage{
rkwarddev.required(min = "0.06-5", lib.loc = NULL)
}
\arguments{
\item{min}{The minimum version number of rkwarddev that is required to run this script.}

\item{lib.loc}{The \code{lib.loc} argument passed over to \code{\link[utils:packageVersion]{packageVersion}}.}
}
\value{
The function has no return value,
      but wil stop with an error if the specified version requirement is not met.
}
\description{
Check for rkwarddev package version requirements
}
\examples{
rkwarddev.required(min="0.06-5")
}

