\name{show}
\alias{show}
\alias{show,-methods}
\alias{show,rk.JS.arr-method}
\alias{show,rk.JS.ite-method}
\alias{show,rk.JS.opt-method}
\title{Show methods for objects of class rk.JS.S}
\arguments{
  \item{object}{An object of class \code{rk.JS.*}}
}
\description{
  Show methods for S4 objects of class \code{rk.JS.*}
}
\keyword{methods}

